Add smr_alignment package
This adds a new top-level package: `smr_alignment`, which will
contain tools for checking for semantic violations of provided
data, as well as aligning said data, in accordance to the SMR
concept.

For now, this adds the `TriangleViolationChecker`, an algorithm
suggested by me for finding and repairing
*Semantic Similarity Triangle Inequality* violations, e.g. edges
in the similarity graph, where: `s(A, C) <= s(A, B) * s(B, C)`.