smr_alignment: Add missing dependency
Previously, the `ecos` solver dependency was missing, somehow not
automatically installed with `cvxpy`. This adds the missing
dependency.

Note, that I had some problems getting `ecos` to install in my
Python 3.13 `venv`, as it is not precompiled for this Python
version. Instead I switched to Python 3.11, like all of my
unittests also already assume.

Fixes #15