smr_alignment.MetricAlignmentOptimizer: Make entropy penalty convex (#19)
Previously, `MetricAlignmentOptimizer.optimize_weights()` used the
`cvxpy` built-in `cvxpy.entr(x)` entropy penalty. What I didn't
realise until I ran the optimizer on real data, is that this
function is actually concave. Therefore, it is possible, that the
overall optimization problem is not convex anymore, messing up our
convex solver.

This adapts the entropy penalty to be a basic L2 penalty,
which is convex.

Fixes #16