algorithm: Fix non-starting-node loops
Previously, there was a bug that potentially caused
non-termination of the algorithm. If there was a loop that did not
include the starting node, and the `min_score` was sufficiently
low, the algorithm could return an arbitrarily large number of
paths going round and round that loop.

While these paths are technically valid matches, they are not
really constructive, in a sense that they do repeat the nodes
that they match, just with different weights. Furthermore, it
is necessary from a technical point of view to avoid them
so that the algorithm terminates.

This adds an additional check that the current path does not run
into such a loop, as well as the corresponding unittest to check
that it works as intended.